+++ src/net/fatekits/lotus/Lotus.java
@@ -1,10 +1,12 @@
 package net.fatekits.lotus;
 
+import com.bizarrealex.azazel.Azazel;
 import net.fatekits.lotus.board.BoardLink;
 import net.fatekits.lotus.colors.ColorGUI;
 import net.fatekits.lotus.colors.ColorListener;
 import net.fatekits.lotus.commands.*;
 import net.fatekits.lotus.configs.LangConfig;
+import net.fatekits.lotus.configs.TabConfig;
 import net.fatekits.lotus.cosmetics.CosmeticInventory;
 import net.fatekits.lotus.cosmetics.CosmeticListener;
 import net.fatekits.lotus.items.ItemListener;
@@ -16,11 +18,14 @@ import net.fatekits.lotus.profiles.ProfileListener;
 import net.fatekits.lotus.profiles.ProfileManager;
 import net.fatekits.lotus.queue.QueueAPI;
 import net.fatekits.lotus.queue.QueueListener;
+import net.fatekits.lotus.ranks.RankAPI;
 import net.fatekits.lotus.ranks.RankManager;
+import net.fatekits.lotus.servers.ServerAPI;
 import net.fatekits.lotus.servers.ServerInventory;
 import net.fatekits.lotus.servers.ServerManager;
 import net.fatekits.lotus.servers.ServerTask;
 import net.fatekits.lotus.settings.SettingsManager;
+import net.fatekits.lotus.tab.TabLink;
 import net.fatekits.lotus.utils.MySQL;
 import net.fatekits.lotus.utils.PluginChannelListener;
 import net.fatekits.lotus.utils.SaveTask;
@@ -29,6 +34,8 @@ import io.github.thatkawaiisam.assemble.AssembleStyle;
 import lombok.Getter;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
+
+import java.io.File;
 import java.util.Arrays;
 import java.util.logging.Logger;
 
@@ -50,14 +57,18 @@ public final class Lotus extends JavaPlugin {
     private QueueAPI queueAPI;
     private ServerTask serverTask;
     private CosmeticInventory cosmeticInventory;
+    private RankAPI rankAPI;
+    private ServerAPI serverAPI;
+    private TabConfig tabConfig;
 
     public void onEnable() {
         plugin = this;
         log.info("[Lotus] initiating...");
         log.info("[Lotus] loading config.yml ...");
-        this.saveConfig();
-        this.saveDefaultConfig();
-        this.reloadConfig();
+        File file = new File(getDataFolder() + "config.yml");
+        if (!file.exists()){
+            saveDefaultConfig();
+        }
         registerConfigs();
         log.info("[Lotus] pre-loading managers");
         registerManagers();
@@ -69,21 +80,27 @@ public final class Lotus extends JavaPlugin {
         registerListeners();
         log.info("[Lotus] pre-loading servers...");
         Lotus.getPlugin().getServerManager().loadServers();
+        log.info("[Lotus] loading ServerAPI...");
+        serverAPI = new ServerAPI();
         log.info("[Lotus] pre-loading all ranks");
         Lotus.getPlugin().getRankManager().loadRanks();
+        log.severe("[Lotus] loading RankAPI...");
+        rankAPI = new RankAPI();
         Bukkit.getMessenger().registerOutgoingPluginChannel(this, "BungeeCord");
         // allow to send to BungeeCord
         Bukkit.getMessenger().registerIncomingPluginChannel(this, "Return", pcl = new PluginChannelListener());
-        log.info("[Lotus] All tasks completed.");
+        log.info("[Lotus] loading scoreboard...");
         Assemble assemble = new Assemble(this, new BoardLink());
         assemble.setTicks(2);
         assemble.setAssembleStyle(AssembleStyle.MODERN);
+        log.info("[Lotus] loading tablist...");
+        new Azazel(this, new TabLink());
+        log.info("[Lotus] All tasks completed.");
     }
 
     public void onDisable() {
-        reloadConfig();
         saveConfig();
-        saveDefaultConfig();
+        reloadConfig();
         mySQL.disconnectProfiles();
         plugin = null;
     }
@@ -91,6 +108,7 @@ public final class Lotus extends JavaPlugin {
     private void registerCommands() {
         getCommand("color").setExecutor(new ColorCommand());
         getCommand("console").setExecutor(new ConsoleCommand());
+        getCommand("items").setExecutor(new ItemsCommand());
         getCommand("leavequeue").setExecutor(new LeaveQueueCommand());
         getCommand("lotus").setExecutor(new LotusCommand());
         getCommand("queue").setExecutor(new QueueCommand());
@@ -116,8 +134,8 @@ public final class Lotus extends JavaPlugin {
     }
 
     private void registerListeners() {
-        Arrays.asList(new ProfileListener(), new ServerInventory(), new ItemListener(), new ColorListener(), new ChatEvent(), new SettingsManager(), new PlayerListener(), new FixEvents(), new QueueListener(),
-                new CosmeticListener())
+        Arrays.asList(new ProfileListener(), new ServerInventory(), new ItemListener(), new ColorListener(), new ChatEvent(), new SettingsManager(), new PlayerListener(), new FixEvents(), new QueueListener()
+                , new CosmeticListener())
                 .forEach(listener -> Bukkit.getPluginManager().registerEvents(listener, this));
     }
 
@@ -128,6 +146,12 @@ public final class Lotus extends JavaPlugin {
         langConfig.saveConfig();
         langConfig.saveDefaultConfig();
         langConfig.reloadConfig();
+        log.info("[Lotus] loading tab.yml ...");
+        tabConfig = new TabConfig(this, "tab.yml");
+        tabConfig.getConfig().options().copyDefaults(true);
+        tabConfig.saveConfig();
+        tabConfig.saveDefaultConfig();
+        tabConfig.reloadConfig();
     }