+++ src/net/fatekits/lotus/colors/ColorListener.java
@@ -19,34 +19,42 @@ public class ColorListener implements Listener {
                 if (event.getClickedInventory() != null) {
                     if (event.getClickedInventory().getName().equalsIgnoreCase(Lotus.getPlugin().getColorGUI().getColorGI().getName())) {
                         if (event.getCurrentItem() != null) {
-                            if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Remove")) {
-                                profile.setColor("&7");
-                                player.sendMessage(StringUtil.format("&cYou have reset your color"));
-                            }
-                            if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Staff-Colors")) {
-                                if (profile.isStaff()) {
-                                    player.openInventory(Lotus.getPlugin().getColorGUI().getStaffColors());
-                                } else {
-                                    player.sendMessage(StringUtil.format("&cYou do not have permission to use these colors."));
+                            if (event.getCurrentItem().getItemMeta() != null) {
+                                if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Remove")) {
+                                    profile.setColor("&7");
+                                    player.sendMessage(StringUtil.format("&cYou have reset your color"));
                                 }
-                            }
-                            if (event.getClickedInventory().getName().equals(Lotus.getPlugin().getColorGUI().getColorGI().getName())) {
-                                event.setCancelled(true);
-                                Color color = Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName());
-                                color.onClick(player, profile, event.getCurrentItem());
+                                if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Staff-Colors")) {
+                                    if (profile.isStaff()) {
+                                        player.openInventory(Lotus.getPlugin().getColorGUI().getStaffColors());
+                                    } else {
+                                        player.sendMessage(StringUtil.format("&cYou do not have permission to use these colors."));
+                                    }
+                                }
+                                if (event.getClickedInventory().getName().equals(Lotus.getPlugin().getColorGUI().getColorGI().getName())) {
+                                    event.setCancelled(true);
+                                    if (Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName()) != null) {
+                                        Color color = Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName());
+                                        color.onClick(player, profile, event.getCurrentItem());
+                                    }
+                                }
                             }
                         }
                     }
                     if (event.getClickedInventory().getName().equalsIgnoreCase(Lotus.getPlugin().getColorGUI().getStaffColors().getName())) {
                         if (event.getCurrentItem() != null) {
-                            if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Remove")) {
-                                profile.setColor("&7");
-                                player.sendMessage(StringUtil.format("&cYou have reset your color"));
-                            }
-                            if (event.getClickedInventory().getName().equals(Lotus.getPlugin().getColorGUI().getColorGI().getName())) {
-                                event.setCancelled(true);
-                                Color color = Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName());
-                                color.onClick(player, profile, event.getCurrentItem());
+                            if (event.getCurrentItem().getItemMeta() != null) {
+                                if (event.getCurrentItem().getItemMeta().getDisplayName().contains("Remove")) {
+                                    profile.setColor("&7");
+                                    player.sendMessage(StringUtil.format("&cYou have reset your color"));
+                                }
+                                if (event.getClickedInventory().getName().equals(Lotus.getPlugin().getColorGUI().getColorGI().getName())) {
+                                    event.setCancelled(true);
+                                    if (Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName()) != null) {
+                                        Color color = Color.getByName(event.getCurrentItem().getItemMeta().getDisplayName());
+                                        color.onClick(player, profile, event.getCurrentItem());
+                                    }
+                                }
                             }
                         }
                     }