+++ src/net/fatekits/lotus/commands/LotusCommand.java
@@ -1,12 +1,14 @@
 package net.fatekits.lotus.commands;
-
 import net.fatekits.lotus.Lotus;
+import net.fatekits.lotus.utils.StringUtil;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 
 public class LotusCommand implements CommandExecutor {
+    FileConfiguration config = Lotus.getPlugin().getLangConfig().getConfig();
 
     @Override
     public boolean onCommand(CommandSender sender, Command command, String s, String[] args) {
@@ -14,10 +16,24 @@ public class LotusCommand implements CommandExecutor {
             return true;
         }
         Player player = (Player) sender;
-        if (player.hasPermission("lotus.rank")) {
-            player.sendMessage("it worked!");
+        if (!player.hasPermission("lotus.admin")) {
+            player.sendMessage(StringUtil.format(config.getString("no-permissions")));
+            return true;
+        }
+        if (args.length == 0) {
+
         } else {
-            Lotus.getPlugin().getRankManager().addPermissions(player, "lotus.rank");
+            if (args.length == 1) {
+                if (args[0].equalsIgnoreCase("reload")) {
+                    Lotus.getPlugin().saveConfig();
+                    Lotus.getPlugin().reloadConfig();
+                    Lotus.getPlugin().getLangConfig().saveConfig();
+                    Lotus.getPlugin().getLangConfig().reloadConfig();
+                    Lotus.getPlugin().getTabConfig().saveConfig();
+                    Lotus.getPlugin().getTabConfig().reloadConfig();
+                    player.sendMessage(StringUtil.format(config.getString("config-reload")));
+                }
+            }
         }
         return false;
     }