+++ src/net/fatekits/lotus/profiles/Profile.java
@@ -2,6 +2,8 @@ package net.fatekits.lotus.profiles;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.bukkit.entity.Player;
+
 import java.util.UUID;
 
 @Getter
@@ -10,8 +12,10 @@ public class Profile {
     private final UUID uuid;
     private String name,IP,rank,color;
     private boolean scoreboard,tablist,chat,visibility,staff;
+    private Player player;
 
-    public Profile(UUID uuid) {
+    public Profile(UUID uuid, Player player) {
         this.uuid = uuid;
+        this.player = player;
     }
 }